+++ mitra_models/src/posts/constants.rs
@@ -1,0 +1,2 @@
+// This search configuration is used in migrations
+pub const PREINSTALLED_FTS_CONFIG: &str = "simple";