+++ mitra_models/src/posts/queries.rs
@@ -1915,11 +1915,14 @@ pub async fn delete_repost(
 
 pub async fn search_posts(
     db_client: &impl DatabaseClient,
+    search_config: &str,
     text: &str,
     current_user_id: Uuid,
     limit: u16,
     offset: u16,
 ) -> Result<Vec<PostDetailed>, DatabaseError> {
+    // `&str` can't be directly cast to `regconfig`
+    // https://github.com/rust-postgres/rust-postgres/issues/1041
     let statement = format!(
         "
         SELECT
@@ -1930,42 +1933,46 @@ pub async fn search_posts(
         JOIN actor_profile ON post.author_id = actor_profile.id
         WHERE
             -- can parse HTML documents
-            to_tsvector('simple', COALESCE(post.title, '') || ' ' || post.content)
-                @@ plainto_tsquery('simple', $1)
+            to_tsvector(
+                $1::text::regconfig,
+                COALESCE(post.title, '') || ' ' || post.content
+            )
+                @@ plainto_tsquery($1::text::regconfig, $2)
             AND repost_of_id IS NULL
             AND (
                 -- posts published by the current user
-                post.author_id = $2
+                post.author_id = $3
                 -- posts bookmarked by the current user
                 OR EXISTS (
                     SELECT 1 FROM bookmark
                     WHERE
                         bookmark.post_id = post.id
-                        AND bookmark.owner_id = $2
+                        AND bookmark.owner_id = $3
                 )
                 -- posts with reactions from the current user
                 OR EXISTS (
                     SELECT 1 FROM post_reaction
                     WHERE
                         post_reaction.post_id = post.id
-                        AND post_reaction.author_id = $2
+                        AND post_reaction.author_id = $3
                 )
                 -- posts where the current user is mentioned
                 OR EXISTS (
                     SELECT 1 FROM post_mention
                     WHERE
                         post_mention.post_id = post.id
-                        AND post_mention.profile_id = $2
+                        AND post_mention.profile_id = $3
                 )
             )
         ORDER BY post.id DESC
-        LIMIT $3 OFFSET $4
+        LIMIT $4 OFFSET $5
         ",
         post_subqueries=post_subqueries(),
     );
     let rows = db_client.query(
         &statement,
         &[
+            &search_config,
             &text,
             &current_user_id,
             &i64::from(limit),
@@ -2022,9 +2029,12 @@ mod tests {
             create_custom_feed,
         },
         database::test_utils::create_test_database,
-        posts::test_utils::{
-            create_test_local_post,
-            create_test_remote_post,
+        posts::{
+            constants::PREINSTALLED_FTS_CONFIG,
+            test_utils::{
+                create_test_local_post,
+                create_test_remote_post,
+            },
         },
         profiles::test_utils::{
             create_test_remote_profile,
@@ -2983,6 +2993,7 @@ mod tests {
         ).await;
         let results = search_posts(
             db_client,
+            PREINSTALLED_FTS_CONFIG,
             "post",
             user.id,
             5,