+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Added `title` parameter to `/api/v1/statuses` form data.
 - Add `title` field to `Status` API entity if ActivityPub object has a `name`.
+- Added `default_fts_config` dynamic configuration parameter for changing text search settings.
 
 ### Changed